Refactor update target tags
